requirements.txt
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
certifi>=2023.7.22
scipy>=1.9.0


